Move QueryDataProducer to a more appropriate package
